Reorganized and corrected some tests
