+++ Tests/TuistCoreTests/Graph/Mappers/ProjectMapperTests.swift
@@ -54,31 +54,6 @@ final class TargetProjectMapperTests: XCTestCase {
         ])
     }
 
-    func test_map_environmentLoggingToggles() throws {
-        // Given
-        let targetMapper = TargetActionEnvironmentMapper(false)
-
-        let subject = TargetProjectMapper(mapper: targetMapper)
-        let scriptA = TargetAction(name: "Pre Script", order: .pre)
-        let scriptB = TargetAction(name: "Post Script", order: .post)
-        let targetA = Target.test(name: "A", actions: [scriptA, scriptB])
-        let targetB = Target.test(name: "B", actions: [scriptA, scriptB])
-        let project = Project.test(targets: [targetA, targetB])
-
-        // When
-        let (updatedProject, _) = try subject.map(project: project)
-
-        // Then
-        XCTAssertTrue(project.targets[0].actions[0].showEnvVarsInLog)
-        XCTAssertTrue(project.targets[0].actions[1].showEnvVarsInLog)
-        XCTAssertTrue(project.targets[1].actions[0].showEnvVarsInLog)
-        XCTAssertTrue(project.targets[1].actions[1].showEnvVarsInLog)
-        XCTAssertFalse(updatedProject.targets[0].actions[0].showEnvVarsInLog)
-        XCTAssertFalse(updatedProject.targets[0].actions[1].showEnvVarsInLog)
-        XCTAssertFalse(updatedProject.targets[1].actions[0].showEnvVarsInLog)
-        XCTAssertFalse(updatedProject.targets[1].actions[1].showEnvVarsInLog)
-    }
-
     // MARK: - Helpers
 
     private class TargetMapper: TargetMapping {
